Lab Session - 6
This Lab session views the SCU MSIS Program through the persona of a recruiter. 
I have worked on highlighting the course coverage of the Top MSIS skills required of the SCU students.
I also have emphasized on leveraging these skills and competencies for future progress of SCU students and where they like to work after graduation.